requirements.txt
flask
tensorflow==2.9.1
pillow
